+++ src/_types/index.d.ts
@@ -1,2 +1,3 @@
 export * from './etsy_listing_image_T'
 export * from './etsy_listing_T'
+export * from './etsy_payload_T'